refactor: removed useless code of icons & unified icon export interface
