+++ src/pages/Content/icon/PlusIcon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-const icon = (
+export default () => (
     <svg
         width="24"
         height="24"
@@ -15,7 +15,3 @@ const icon = (
         <line x1="8" y1="12" x2="16" y2="12" />
     </svg>
 );
-
-export function PlusIcon() {
-    return icon;
-}
