+++ src/pages/Content/icon/index.tsx
@@ -1,0 +1,6 @@
+import PlusIcon from "./PlusIcon";
+import NoteIcon from "./NoteIcon";
+import SwitchIcon from "./SwitchIcon";
+import HomeIcon from "./HomeIcon";
+
+export { PlusIcon, NoteIcon, SwitchIcon, HomeIcon };
